HUDI 822 decouple Hudi related logics from HoodieInputFormat
